remove unecessary and slow previous items fetching
